ref(*) Use fewer interfaces.
